KAN-97 style: 폴더 구조 변경
